+++ js/data.js
@@ -1,0 +1,65 @@
+//  object author
+
+const AVATAR = [
+  'img/avatars/user01.png',
+  'img/avatars/user02.png',
+  'img/avatars/user03.png',
+  'img/avatars/user04.png',
+  'img/avatars/user05.png',
+  'img/avatars/user06.png',
+  'img/avatars/user07.png',
+  'img/avatars/user08.png',
+];
+
+//  object offer
+
+const TITLE = [
+  'Аппартаменты на любой вкус!',
+  'Вам понравится!',
+  'Бронируйте, не пожалеете!',
+  'У нас печеньки!',
+];
+const TYPE = [
+  'palace',
+  'flat',
+  'house',
+  'bungalow',
+];
+const CHECKIN = [
+  '12:00',
+  '13:00',
+  '14:00',
+];
+const CHECKOUT = [
+  '12:00',
+  '13:00',
+  '14:00',
+];
+const FEATURES = [
+  'wifi',
+  'dishwasher',
+  'parking',
+  'washer',
+  'elevator',
+  'conditioner',
+];
+const DESCRIPTION = [
+  'В каждом номере бар',
+  'Душ с горячей водой',
+  'Номер для некурящих',
+  'Джакузи в номере',
+];
+const PHOTOS = [
+  'http://o0.github.io/assets/images/tokyo/hotel1.jpg',
+  'http://o0.github.io/assets/images/tokyo/hotel2.jpg',
+  'http://o0.github.io/assets/images/tokyo/hotel3.jpg',
+];
+
+//  object location
+
+//  число с плавающей точкой — широта, случайное значение от 35.65000 до 35.70000
+//  число с плавающей точкой — долгота, случайное значение от 139.70000 до 139.80000
+
+const ADVERTISEMENT_NUMBER = 10;
+
+export {AVATAR, TITLE, TYPE, CHECKIN, CHECKOUT, FEATURES, DESCRIPTION, PHOTOS, ADVERTISEMENT_NUMBER};
